requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
datatig==0.8.0
    # via
    #   -r requirements.in
    #   staticpipes-datatig
dateparser==1.2.1
    # via datatig
feedgen==0.9.0
    # via -r requirements.in
jinja2==3.0.3
    # via
    #   datatig
    #   staticpipes
jsonschema==4.23.0
    # via datatig
jsonschema-specifications==2025.4.1
    # via jsonschema
lxml==5.4.0
    # via feedgen
markdown-it-py==3.0.0
    # via
    #   -r requirements.in
    #   staticpipes
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
pygments==2.19.1
    # via datatig
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   feedgen
pytz==2025.2
    # via
    #   datatig
    #   dateparser
pyyaml==6.0.2
    # via
    #   datatig
    #   staticpipes
rcssmin==1.2.1
    # via staticpipes
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via dateparser
rjsmin==1.2.4
    # via staticpipes
rpds-py==0.25.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
staticpipes[allbuild]==0.6.0
    # via
    #   -r requirements.in
    #   staticpipes-datatig
staticpipes-datatig @ git+https://github.com/TeacakeTech/StaticPipes-DataTig.git@7dc9a28b04580386e8c26a00099592bf4138dd3c
    # via -r requirements.in
tzlocal==5.3.1
    # via dateparser
